+++ lib/leftovers/file_collector/node_processor/error.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Leftovers
+  class FileCollector
+    class NodeProcessor
+      class Error < ::Leftovers::FileCollector::Error
+        attr_reader :node
+
+        def initialize(message, node)
+          super(message)
+          @node = node
+        end
+      end
+    end
+  end
+end